// packages/types/src/generated/prisma/models/comments.ts
export type commentsSelect<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetSelect<{
  id?: boolean
  work_item_id?: boolean
  author_id?: boolean
  parent_id?: boolean
  content?: boolean
  edited?: boolean
  status?: boolean
  created_by?: boolean
  created_at?: boolean
  updated_by?: boolean
  updated_at?: boolean
  work_item?: boolean | Prisma.work_itemsDefaultArgs<ExtArgs>
  author?: boolean | Prisma.usersDefaultArgs<ExtArgs>
  parent?: boolean | Prisma.comments$parentArgs<ExtArgs>
  replies?: boolean | Prisma.comments$repliesArgs<ExtArgs>
  created_by_user?: boolean | Prisma.comments$created_by_userArgs<ExtArgs>
  updated_by_user?: boolean | Prisma.comments$updated_by_userArgs<ExtArgs>
  _count?: boolean | Prisma.CommentsCountOutputTypeDefaultArgs<ExtArgs>
}, ExtArgs["result"]["comments"]>
// ...
export type commentsOmit<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetOmit<"id" | "work_item_id" | "author_id" | "parent_id" | "content" | "edited" | "status" | "created_by" | "created_at" | "updated_by" | "updated_at", ExtArgs["result"]["comments"]>
export type commentsInclude<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  work_item?: boolean | Prisma.work_itemsDefaultArgs<ExtArgs>
  author?: boolean | Prisma.usersDefaultArgs<ExtArgs>
  parent?: boolean | Prisma.comments$parentArgs<ExtArgs>
  replies?: boolean | Prisma.comments$repliesArgs<ExtArgs>
  created_by_user?: boolean | Prisma.comments$created_by_userArgs<ExtArgs>
  updated_by_user?: boolean | Prisma.comments$updated_by_userArgs<ExtArgs>
  _count?: boolean | Prisma.CommentsCountOutputTypeDefaultArgs<ExtArgs>
}
// ...
export type $commentsPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "comments"
  objects: {
    work_item: Prisma.$work_itemsPayload<ExtArgs>
    author: Prisma.$usersPayload<ExtArgs>
    parent: Prisma.$commentsPayload<ExtArgs> | null
    replies: Prisma.$commentsPayload<ExtArgs>[]
    created_by_user: Prisma.$usersPayload<ExtArgs> | null
    updated_by_user: Prisma.$usersPayload<ExtArgs> | null
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    work_item_id: string
    author_id: string
    parent_id: string | null
    content: runtime.JsonValue
    edited: boolean
    status: $Enums.RecordStatus
    created_by: string | null
    created_at: Date
    updated_by: string | null
    updated_at: Date
  }, ExtArgs["result"]["comments"]>
  composites: {}
}

export type commentsGetPayload<S extends boolean | null | undefined | commentsDefaultArgs> = runtime.Types.Result.GetResult<Prisma.$commentsPayload, S>
// ...
export type commentsDefaultArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the comments
   */
  select?: Prisma.commentsSelect<ExtArgs> | null
  /**
   * Omit specific fields from the comments
   */
  omit?: Prisma.commentsOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.commentsInclude<ExtArgs> | null
}